# Conflicts:
#	app/src/main/java/com/practicum/spisokpokupok/lists/presentation/currentlists/CurrentPurchasesListScreen.kt

+++ app/src/main/java/com/practicum/spisokpokupok/core/data/roomdb/mapper/ExternalToLocalMapper.kt
@@ -5,6 +5,7 @@ import com.practicum.spisokpokupok.core.data.roomdb.entity.LocalShoppingList
 import com.practicum.spisokpokupok.core.data.roomdb.entity.LocalShoppingTask
 import com.practicum.spisokpokupok.listdetails.domain.model.Good
 import com.practicum.spisokpokupok.listdetails.domain.model.Task
+import com.practicum.spisokpokupok.listdetails.domain.model.quantityTypeToString
 import com.practicum.spisokpokupok.lists.domain.model.ShoppingList
 
 fun Good.toLocal() =
@@ -17,7 +18,7 @@ fun Task.toLocal(
     localGoodId: String,
     shoppingListId: String,
 ) = LocalShoppingTask(
-    id = id.toInt(),
+    id = id,
     goodId = localGoodId.toInt(),
     isCompleted = isCompleted,
     shoppingListId = shoppingListId,
